Fetch blocks only one time for all generated reports
Instead of fetching the same blocks for each customer the report is
generated for, we now fetch all blocks upfront and then filter out
transactions for each customer based on blocks cached in the memory.

Fetching blocks is the most time-consuming activity during generation of
reports so this change speeds up the process significantly.

I have also added some additional logs informing about the progress of
getting blocks as well as added an initite retry with 60 sec sleep for
errors occurred during getting the blocks. We don't want the activity
taking such a long time to stop in the middle because of some single
failure.