Changed the algorithm for computing staker rewards
All ETH on beneficiary account goes now to the staker. The algorithm
does no longer compute operational costs.

There is a couple reasons for that. First and foremost, some operators
use the same account for ECDSA and beacon clients. Computing operational
costs of beacon is not possible.

Also, this approach is more staker-friendly. If they were lucky to with
1st group member seat and they receive extra submitter rewards, that's
great for them, and let's forward all that ETH to the staker without
taking a share from it.