A mix of fixes for how group counts are calculated
This commit should be split into three separate commits but... it is
hard to fix one thing without fixing another and we are where we are,
heh.

The algorithm for calculating the number of active, expired, and total
groups created in the system is fixed.

We now use function from random beacon operator contract to get the
total number of created groups which works fine even if there are some
expired groups in the system.